requirements.txt
umap-learn
joblib
scipy
pandas
numpy
matplotlib
seaborn
scikit-learn
missingno
xgboost
shap
torch